requirements.txt
absl-py==0.6.1
astor==0.7.1
backports.weakref==1.0.post1
bleach==1.5.0
boto==2.49.0
boto3==1.9.47
botocore==1.12.47
bz2file==0.98
certifi==2018.10.15
chardet==3.0.4
cloudpickle==0.6.1
cycler==0.10.0
dask==0.20.2
decorator==4.3.0
docutils==0.14
enum34==1.1.6
funcsigs==1.0.2
functools32==3.2.3.post2
future==0.17.1
futures==3.2.0
gast==0.2.0
gensim==3.6.0
grpcio==1.16.1
gym==0.10.8
h5py==2.8.0
hmmlearn==0.2.1
html5lib==0.9999999
idna==2.7
jmespath==0.9.3
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
Markdown==3.0.1
matplotlib==2.0.0
mock==2.0.0
networkx==2.2
numpy==1.15.4
pbr==5.1.1
Pillow==5.3.0
protobuf==3.6.1
pyglet==1.3.2
PyOpenGL==3.1.0
pyparsing==2.3.0
python-dateutil==3.5
pytz==2018.7
PyWavelets==1.0.1
requests==2.20.0
s3transfer==0.1.13
scikit-image==0.14.1
scikit-learn==0.20.0
scipy==1.1.0
six==1.11.0
sklearn==0.0
smart-open==1.7.1
subprocess32==3.5.3
tensorboard==1.12.0
tensorflow-gpu==1.12.0
tensorflow-tensorboard==0.4.0
termcolor==1.1.0
tflearn==0.3.2
toolz==0.9.0
urllib3==1.24.1
Werkzeug==0.14.1
xgboost==0.80